Switch from flat files to PostgreSQL.
The flat files don't work all that well when you have multiple clients
working on them at the same time. I'm not super-happy about requiring
a server to be up all the time (it would seem to reduce resilience
by quite a lot), but Postgres is a pretty stable thing, so hopefully
we will manage.

Include a script to ingest old data into the database, which we will
remove immediately afterwards.